Remove `.json` extension from URL paths
According to the swagger definition, this API should only give JSON
responses and does not require a `.json` extension to be added to
requests to endpoints.

This commit removes the requirement for callers to specify `.json` by
configuring the router to default to JSON format.

Now, every call to endpoint:
```
/complexity-of-need/offender-no/:offender_no
```

will respond with JSON, even if the `.json` extension isn't added.