Add versioning and server URLs to Swagger doc
This commit add a `/v1` prefix to all endpoint paths.

Adding versioning to our API paths upfront gives us the ability to
accommodate future breaking changes – i.e. by creating a `v2`, and so on

It also adds the expected hostnames for the API – although these aren't
yet setup in the Cloud Platform.